Add/consolidate intrinsic and dereference mutability tests
Added a number of tests for testing intrinsic and dereference
mutability through operations like assignment, &, *, x.y, and x[y].
Dereference mutability of left had to be at least as weak as intrinsic
mutability of right in assignment of types with pointers.  This was a
bug and removed.  The & operator did not change the intrinsic and
dereference mutability of the lvalue.  This was a bug.  Now, the
deference mutability of the pointer is the weaker of the intrinsic and
dereference mutability of the lvalue.