Remove check that +foreign values are not stored to heap
Suppose we have x = y where y has contains pointers and is declared
+foreign.  What we don't want is for x to refer to a location on the
heap.  However, the only way to generate an lvalue that refers to the
heap is to dereference a pointer or index a slice so the assignment is
either *x = y or x[z] = y.  Since the RHS is +foreign, the LHS must
also be +foreign.  If *x or x[z] is +foreign, then x must be +foreign.
If x is +foreign, then *x and x[z] must be foreign.  Thus, the LHS of
the assignment is not a valid.  The explicit check that the LHS was on
the stack was removed as it is never reached.