Consolidate function/method types, memoize derived types
The Type namespace was introduced and types were renamed.  Primitive
types (like bool, int, etc.) were rewritten using a template.
Pointers, slices, arrays, and heaps are memoized as they are needed to
avoid creating duplicates.  The port types, initializers, and getters
were folded into Function and Method.  The goal of these changes is to
reduce the amount of code related to type descriptors.

I would like to clean up the functions in Type.  Many of them should
be converted to methods.  Similarly, the Visitor class should be
reconsidered.  It seems like a default implementation of the visit
methods is the wrong thing to do.  Defining different visitors for
different classes of type descriptors might be the way to go.  For
example, one visitor might only handle arithmetic types.  That is, it
would force overrides for the arithmetic types and provide a default
(error) for the other types.