Fix composition check for recursive graphs
A general overhaul of the composition checking code.  The checking
code is now much more graph oriented.  The nodes in the graph
correspond to enumerated actions, reactions, getters, push ports, pull
ports, and activations.  (Something is enumerated if it is associated
with a particular instances.)  The check now checks for recursive
structure (which is illegal) and handles direct calls to getters.