Fix next block size
The next block size was only reset if all blocks were collected for
the heap.  Suppose that one block is retained but is completely
allocated.  The next allocation will allocate a new block because
nothing is free which will increase the next block size.  After
collection, the scenario starts again but with a larger next block
size.  Thus, the block size was increasing without bound.

The solution was to reset the next block size after collection.
Currently, it is equal to the allocated size of the heap.