Fix composition analysis for getters in preconditions
A precondition may call a getter.  The instance set for an action
should include instances accessed via getter in the precondition.
This change extends the instance set calculation to the precondition
of actions.

Similarly, an action whose precondition contains a getter may become
enabled when the dependent component changes state.  The composition
analysis was updated to compute the set of instances that may become
enabled when an action in an instance is executed.  That is, given
that instance I changed state, the set D represents all the components
with preconditions that depend on I.  Thus, all the components in D
should be added to the schedule when I changes.  The instance
scheduler was updated to process D.  Changing the partitioned
scheduler was not necessary.